(slime-net-connect, slime-net-send): Set coding system to 'no-conversion and convert mutlibyte-strings to unibyte before sending.
(slime-process-available-input): Read all available input, even when
someone throws out of slime-take-input.

(slime-eval): Invoke the debugger on the Lisp and Emacs side when an
error occurs on the Lisp side.  We do this because the Emacs debugger
runs in a recursive edit.  When the Lisp debugger aborts or continues
then leave the Emacs debugger and vice versa.

(slime-space): Show the arglist only when we are in idle-state or
sldb-state.

(slime-show-apropos): Turn on slime minor mode, so we can use M-. here.

(slime-describe-symbol): Don't query by default.

(slime-interrupt, slime-quit): New commands.  Only available in idle-
and sldb- states.

(sldb-toggle-details): New commands.  Show local variables and
catch-tags in the backtrace.