(slime-display-message-or-view, slime-remove-message-window): Also display too long lines in a new window.  Add a temporary pre-command-hook to remove the multiline window before the next command is executed. (slime-display-buffer-region): Some of the comments where out of sync with the code.
(slime-complete-symbol): Save the window configuration before
displaying the completions and try to restore it later.  The
configuration is restored when: (a) the completion is unique (b) there
are no completion.  It is also possible to delay the restoration until
(c) certain characters, e.g, space or a closing paren, are inserted.

(slime-selector): Don't abort when an unkown character is pressed;
display a message and continue.  Similiar for ?\?.  Add a selector for
the *sldb* buffer.

(slbd-hook, sldb-xemacs-post-command-hook): Emulate Emacs'
point-entered text property with a post-command hook.