(slime-multiprocessing): When true, use multiprocessing in Lisp if available.
(slime-global-debugger-hook): When true, globally set *debugger-hook*
to use the SLIME debugger. For use with SERVE-EVENT and
multiprocessing.

(slime-handle-oob): Handle :AWAITING-GOAHEAD message from threads that
have suspended to wait for Emacs's attention.

(slime-give-goahead): New command to allow a suspended thread to
continue (bound to RET in the thread-control-panel).

(slime-thread-control-panel): New command to display a buffer showing
all threads that are suspending waiting for Emacs's attention. Bound
to `C-c C-x t'.

(slime-popup-thread-control-panel): When true, automatically popup the
thread-control buffer when a new thread suspends.