(slime-easy-menu): Add some commands.
(slime-changelog-date): New variable. Initialized with the value
returned by the function of the same name.  This detects incompatible
versions if Emacs has not been restarted after an upgrade.
(slime-check-protocol-version, slime-init-output-buffer): Use it.

(slime-events-buffer, slime-log-event): Use fundamental mode instead
of lisp-mode to avoid excessive font-locking for messages with lots of
strings.