(slime-goto-source-location): Added support for the :snippet "hint" in a location specifier. If Lisp sends the (initial) source text for the definition then Emacs isearches for it in both directions from the given character position. This makes M-. robust when the Emacs buffer has been edited. Requires backends to provide this snippet information.
(slime-goto-location-position): Tightened up the regular expressions
for :function-name style location search.

(slime-cleanup-definition-refs): New function to do a little
post-processing on definition references from Lisp. Mostly this is a
hack: if POSITION is NIL then we fill it in with the function name,
ready for regexp search. I was in a hurry and it was easier to do
here, and it doesn't seem entirely unreasonable.