(*sldb-pprint-frames*): Renamed to *sldb-print-pretty*.
(*sldb-print-level*, *sldb-print-length*, *sldb-print-circle*)
(*sldb-print-readbly): Group of new variables to customize printing in
the debugger.  The default values should be safe.

(define-printer-variables, with-printer-settings): New macros to make
definig and binding groups printer variables easier.

(inspect-for-emacs-list): Rewritten. The old version had a bug with
circular lists, didn't include the position of the element, and always
showed the full list.  The new version only shows the first 40
elements.

(inspect-for-emacs): Minor cleanups.

(all-qualified-readnames): Removed. It was not needed because
common-lisp-indent-function strips of any package prefix and downcases
the symbol anyway.